statefiles: move dhcpv6_ia_enum_addrs() to odhcpd.c
dhcpv6_ia_enum_addrs() is used in several different places (ubus,
statesfile, dhcpv6-ia), so move it to a more central location. At the
same time, rename it to odhcpd_enum_addr6().

Link: https://github.com/openwrt/odhcpd/pull/302
